+++ src/consts.js
@@ -30,3 +30,9 @@ export const MenuItem = {
   TABLE: `TABLE`,
   STATISTICS: `STATISTICS`
 };
+
+export const State = {
+  SAVING: `SAVING`,
+  DELETING: `DELETING`,
+  ABORTING: `ABORTING`
+};
